pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
exclude = ["tests*", "docs*", "notebooks*", "frontend*", "datasets*", "scripts*"]

[tool.setuptools]
py-modules = []

[project]
name = "fx-alphalab"
version = "0.1.0"
description = "Multi-Agent AI Framework for FX Market Analysis"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "FX-AlphaLab Team"}
]

dependencies = [
    "pandas",
    "numpy",
    "statsmodels",
    "python-dotenv",
    "psycopg2-binary>=2.9",
    "requests",
    "beautifulsoup4",
    "feedparser",
    "fredapi",
    "feedparser",
    "sqlalchemy",
    "alembic>=1.13",
    "google-api-core",
    "pytz",
    "pyarrow",
    "pytrends",
    "db-dtypes",
    "lxml",
    "selenium>=4.0",
    "undetected-chromedriver>=3.5",
    "webdriver-manager>=4.0",
    "transformers>=4.30",
    "torch>=2.0",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "apscheduler>=3.10",
]


[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "ruff",
    "mypy",
    "pre-commit",
    "ipykernel",
    "google-cloud-bigquery>=3.17.0",
    "types-pytz",
]

notebook = [
    "jupyter",
    "matplotlib",
    "seaborn",
    "scipy",
    "plotly",
    "yfinance",
]

mt5 = [
    "MetaTrader5",  # Windows only - for MT5 data ingestion
]

[tool.black]
line-length = 100
target-version = ['py310']
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    "*.ipynb",
]
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]

[tool.ruff.per-file-ignores]
"scripts/*" = ["E402"]
"src/ingestion/preprocessors/macro_signal_builder.py" = ["N806"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=. --cov-report=term-missing"
markers = [
    "integration: requires real data files on disk (signals_aligned.parquet, OHLCV parquets)",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
